requirements.txt
bcrypt==3.1.4
cffi==1.11.2
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Script==2.0.6
Frozen-Flask==0.15
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
pathspec==0.5.5
pycparser==2.18
python-frontmatter==0.4.2
python-slugify==1.2.4
PyYAML==3.12
six==1.11.0
Unidecode==0.4.21
Werkzeug==0.12.2
